requirements.txt
# ============================
# Core
# ============================
Django==5.2.7
djangorestframework==3.15.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
python-dotenv==1.2.1

# ============================
# Database
# ============================
mysqlclient==2.2.7

# ============================
# Background Tasks
# ============================
celery==5.5.3
django-celery-beat==2.8.1
redis==7.0.1

# ============================
# AI & ML
# ============================
openai==2.7.1
langdetect==1.0.9
pytesseract==0.3.13

# ============================
# Google Cloud / Firebase
# ============================
firebase-admin==7.1.0
google-cloud-vision==3.11.0
google-cloud-storage==3.5.0
google-cloud-firestore==2.21.0
pyfcm==2.1.0

# ============================
# File Processing
# ============================
PyPDF2==3.0.1
pdf2image==1.17.0
python-docx==1.2.0
python-pptx==1.0.2
xlsxwriter==3.2.9
Pillow==10.3.0

# ============================
# Utilities
# ============================
requests==2.32.5
pytz==2024.1
tqdm==4.67.1

# ============================
# Production server
# ============================
gunicorn
